Made the command to generate a SECRET_KEY_BASE more portable
macOS treats input files to tr as UTF-8 by default and the chances of /dev/random spitting out valid UTF-8 is low (which causes tr to fail and output nothing). Also, the complement of "A-Za-z0-9" seems to allow for invalid characters for some reason, but the alnum character class doesn't.